% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_teams_season.R
\name{all_teams_season}
\alias{all_teams_season}
\title{Return all teams player statistics for a season}
\usage{
all_teams_season(season)
}
\arguments{
\item{season}{A number. 2013 - 2018}
}
\value{
A dataframe, all player game statistics, all teams
}
\description{
Return all teams player statistics for a season
}
