% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_season.R
\name{team_season}
\alias{team_season}
\title{Retrieve team player statistics for a particular season}
\usage{
team_season(team, season)
}
\arguments{
\item{team}{A number.}

\item{season}{A number. 2013 - 2018}
}
\value{
A dataframe, all player game statistics
}
\description{
Retrieve team player statistics for a particular season
}
